renamed scss module
