+++ src/renderer/components/Player/PlayerStyles.module.scss
